Move trace data code to separate file
