+++ crates/cheatnet/src/lib.rs
@@ -6,3 +6,4 @@ pub mod predeployment;
 pub mod runtime_extensions;
 pub mod state;
 pub mod sync_client;
+pub mod trace_data;
